+++ tests/WaitsForElementsTest.php
@@ -2,11 +2,11 @@
 
 namespace Laravel\Dusk\Tests;
 
-use stdClass;
-use Mockery as m;
-use Laravel\Dusk\Browser;
-use PHPUnit\Framework\TestCase;
 use Facebook\WebDriver\Exception\TimeOutException;
+use Laravel\Dusk\Browser;
+use Mockery as m;
+use PHPUnit\Framework\TestCase;
+use stdClass;
 
 class WaitsForElementsTest extends TestCase
 {
@@ -93,6 +93,19 @@ class WaitsForElementsTest extends TestCase
         $browser = new Browser(new stdClass, $resolver);
 
         $browser->waitForText('Discount: 20%');
+    }
+
+    public function test_can_wait_for_text_to_go_missing()
+    {
+        $element = m::mock(stdClass::class);
+        $element->shouldReceive('getText')
+            ->times(3)
+            ->andReturn('Discount: 20%', 'Discount: 20%', 'SOLD OUT!');
+        $resolver = m::mock(stdClass::class);
+        $resolver->shouldReceive('findOrFail')->with('')->andReturn($element);
+        $browser = new Browser(new stdClass, $resolver);
+
+        $browser->waitUntilMissingText('Discount: 20%');
     }
 
     public function test_wait_for_text_failure_message_containing_a_percent_character()