+++ CHANGELOG.md
@@ -1,12 +1,133 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/dusk/compare/v5.1.0...5.0)
+## [Unreleased](https://github.com/laravel/dusk/compare/v5.9.0...master)
+
+### Changed
+- Dropped support for Laravel 5.8 ([e0c173a](https://github.com/laravel/dusk/commit/e0c173af6deaeda6170e017eb3ce225d8d4c5964))
+- Dropped support for PHP 7.1 ([210e03e](https://github.com/laravel/dusk/commit/210e03ec2c121517b99bd6163859bbdc5cce564a))
+- Bumped minimum Symfony dependencies to 4.3 ([1ee28e1](https://github.com/laravel/dusk/commit/1ee28e1bfcce1de4a3ad83253394d964690602c4))
+
+
+## [v5.9.0 (2020-01-28)](https://github.com/laravel/dusk/compare/v5.8.2...v5.9.0)
+
+### Added
+- Add `clickAtXPath` ([#723](https://github.com/laravel/dusk/pull/723), [effe73d](https://github.com/laravel/dusk/commit/effe73d6eb61b4bd77f88814bcd679e4fceb6f25))
+- Add `ProvidesBrowser::getCallerName()` ([#725](https://github.com/laravel/dusk/pull/725))
+
+### Fixed
+- Fit content to `<html>` instead of `<body>` ([#726](https://github.com/laravel/dusk/pull/726))
+
+
+## [v5.8.2 (2020-01-21)](https://github.com/laravel/dusk/compare/v5.8.1...v5.8.2)
+
+### Changed
+- Rename php-webdriver package ([#720](https://github.com/laravel/dusk/pull/720))
+- Update jQuery file ([#721](https://github.com/laravel/dusk/pull/721))
+
+### Fixed
+- Update `fitContent()` ([#717](https://github.com/laravel/dusk/pull/717))
+
+
+## [v5.8.1 (2020-01-07)](https://github.com/laravel/dusk/compare/v5.8.0...v5.8.1)
+
+### Fixed
+- Cast boolean values to appropriate string ([#713](https://github.com/laravel/dusk/pull/713))
+
+
+## [v5.8.0 (2019-12-30)](https://github.com/laravel/dusk/compare/v5.7.0...v5.8.0)
+
+### Added
+- Add "waitUntilMissingText" ([#706](https://github.com/laravel/dusk/pull/706))
+- Add ability to store source from browser ([#707](https://github.com/laravel/dusk/pull/707), [9c90e2a](https://github.com/laravel/dusk/commit/9c90e2a716030c9a36e6306c3f67d606a254bbb7), [1d5bc20](https://github.com/laravel/dusk/commit/1d5bc203b67ffc5a17eb1b89f3e22547e3ea174b))
+
+
+## [v5.7.0 (2019-12-17)](https://github.com/laravel/dusk/compare/v5.6.3...v5.7.0)
+
+### Added
+- Automatically fit content on failures ([#704](https://github.com/laravel/dusk/pull/704))
+
+
+## [v5.6.3 (2019-12-03)](https://github.com/laravel/dusk/compare/v5.6.2...v5.6.3)
+
+### Added
+- Support phpdotenv v4 ([#699](https://github.com/laravel/dusk/pull/699))
+
+### Fixed
+- scrollTo: add support for selectors with quotes ([#697](https://github.com/laravel/dusk/pull/697))
+
+
+## [v5.6.2 (2019-11-26)](https://github.com/laravel/dusk/compare/v5.6.1...v5.6.2)
+
+### Changed
+- Allow for Symfony 5 ([#696](https://github.com/laravel/dusk/pull/696))
+
+
+## [v5.6.1 (2019-11-12)](https://github.com/laravel/dusk/compare/v5.6.0...v5.6.1)
+
+### Fixed
+- Ensure jQuery for scrollTo ([#686](https://github.com/laravel/dusk/pull/686))
+- Added missing return statement in withDuskEnvironment ([#691](https://github.com/laravel/dusk/pull/691))
+- Prevent using pcntl when not installed ([#692](https://github.com/laravel/dusk/pull/692))
+
+
+## [v5.6.0 (2019-10-29)](https://github.com/laravel/dusk/compare/v5.5.2...v5.6.0)
+
+### Added
+- Add scrollTo method ([#684](https://github.com/laravel/dusk/pull/684))
+
+### Fixed
+- Add graceful handler for `SIGINT` for .env restoration ([#682](https://github.com/laravel/dusk/pull/682), [f843b8a](https://github.com/laravel/dusk/commit/f843b8a51ae96933cefcc74dec515377d3135611))
+
+
+## [v5.5.2 (2019-09-24)](https://github.com/laravel/dusk/compare/v5.5.1...v5.5.2)
+
+### Fixed
+- Improve detection of latest stable ChromeDriver release ([#677](https://github.com/laravel/dusk/pull/677))
+
+
+## [v5.5.1 (2019-09-12)](https://github.com/laravel/dusk/compare/v5.5.0...v5.5.1)
+
+### Fixed
+- Update regular expression base on website changes ([#674](https://github.com/laravel/dusk/pull/674))
+
+
+## [v5.5.0 (2019-08-06)](https://github.com/laravel/dusk/compare/v5.4.0...v5.5.0)
+
+### Added
+- Allow saving screenshots in a subdirectory ([#662](https://github.com/laravel/dusk/pull/662))
+
+
+## [v5.4.0 (2019-07-30)](https://github.com/laravel/dusk/compare/v5.3.0...v5.4.0)
+
+### Added
+- Add assertion checks if a button is disabled or enabled ([#661](https://github.com/laravel/dusk/pull/661))
+
+### Fixed
+- Update constraints for Laravel 6.0 ([e4b4d63](https://github.com/laravel/dusk/commit/e4b4d63c179bb1f228db22852bd776db750d1ec6))
+
+
+## [v5.3.0 (2019-07-11)](https://github.com/laravel/dusk/compare/v5.2.0...v5.3.0)
+
+### Added
+- Add proxy support to the `dusk:install` command ([#659](https://github.com/laravel/dusk/pull/659))
+
+
+## [v5.2.0 (2019-06-25)](https://github.com/laravel/dusk/compare/v5.1.1...v5.2.0)
+
+### Added
+- Add option to fullsize the browser ([#655](https://github.com/laravel/dusk/pull/655))
+
+
+## [v5.1.1 (2019-06-14)](https://github.com/laravel/dusk/compare/v5.1.0...v5.1.1)
+
+### Fixed
+- Add latest version of Facebook Webdriver ([#654](https://github.com/laravel/dusk/pull/654))
 
 
 ## [v5.1.0 (2019-05-02)](https://github.com/laravel/dusk/compare/v5.0.3...v5.1.0)
 
 ### Added
-- Implement ChromeDriverCommand ([#643](https://github.com/laravel/dusk/pull/643), [60339a5](https://github.com/laravel/dusk/commit/60339a521a1b05e55af7c90b3151557100a0db4d), [#644](https://github.com/laravel/dusk/pull/644)) 
+- Implement ChromeDriverCommand ([#643](https://github.com/laravel/dusk/pull/643), [60339a5](https://github.com/laravel/dusk/commit/60339a521a1b05e55af7c90b3151557100a0db4d), [#644](https://github.com/laravel/dusk/pull/644))
 
 
 ## [v5.0.3 (2019-04-02)](https://github.com/laravel/dusk/compare/v5.0.2...v5.0.3)